armv8: add arch-pxa1928 support
add arch-pxa1928 support under arch/arm/cpu/armv8/pxa1928
arch/arm/include/asm/arch-pxa1928/

ported from m2011.09:
1. arch/arm/cpu/armv8/pxa1928/cpu.c:
1.1. removed CONFIG_PXA168_FB/I2C_MV related code
1.2. kept I2C_MV/SDHCI/USB clock enabling of APMU
2. arch/arm/cpu/armv8/pxa1928/smp_init.S: removed due to:
2.1. release slave cpu and set release addr is no need for mcpm or
psci and might be harmful.
2.2. LPM test is removed.
3. arch/arm/include/asm/arch-pxa1928/cpu.h
3.1. removed cpu_is_ca7/9
3.2. removed sdh_ functions
4. arch/arm/include/asm/arch-pxa1928/config.h
4.1. removed configures for i2c/usb/mmc/lcd

added:
1. arch/arm/include/asm/arch-pxa1928/cpu.h
1.1. extern smp_config/smp_hw_cpuid out so that no need to extern in c

+++ drivers/mmc/mv_sdhci.c
@@ -30,7 +30,7 @@ static inline void mv_sdhci_writeb(struct sdhci_host *host, u8 val, int reg)
 #endif /* CONFIG_MMC_SDHCI_IO_ACCESSORS */
 
 static char *MVSDH_NAME = "mv_sdh";
-int mv_sdh_init(u32 regbase, u32 max_clk, u32 min_clk, u32 quirks)
+int mv_sdh_init(unsigned long regbase, u32 max_clk, u32 min_clk, u32 quirks)
 {
 	struct sdhci_host *host = NULL;
 	host = (struct sdhci_host *)malloc(sizeof(struct sdhci_host));
